Use plain reflection instead of Roslyn for annotation tests
At some point we may want to revisit Roslyn, but it's not worth it right now - only one test really needed it.
NCrunch, testing under Mono, and testing under DNX all suffered due to this.
